Remove some api_common dependencies (ref #5642) (#5678)
