+++ crates/api/src/local_user/list_saved.rs
@@ -1,14 +1,12 @@
 use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
-use lemmy_api_common::{
-  context::LemmyContext,
-  person::{ListPersonSaved, ListPersonSavedResponse},
-  utils::check_private_instance,
-};
+use lemmy_api_common::{context::LemmyContext, utils::check_private_instance};
 use lemmy_db_schema::traits::PaginationCursorBuilder;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_db_views_person_saved_combined::{
   impls::PersonSavedCombinedQuery,
+  ListPersonSaved,
+  ListPersonSavedResponse,
   PersonSavedCombinedView,
 };
 use lemmy_db_views_site::SiteView;
