+++ crates/api_common/Cargo.toml
@@ -54,11 +54,7 @@ full = [
   "lemmy_db_views_reports/full",
   "lemmy_db_views_private_message/full",
   "lemmy_db_views_local_image/full",
-  "lemmy_db_views_inbox_combined/full",
-  "lemmy_db_views_person_content_combined/full",
-  "lemmy_db_views_person_saved_combined/full",
   "lemmy_db_views_report_combined/full",
-  "lemmy_db_views_search_combined/full",
   "lemmy_db_views_site/full",
   "lemmy_db_views_registration_applications/full",
 ]
@@ -77,11 +73,7 @@ lemmy_db_views_custom_emoji = { workspace = true }
 lemmy_db_views_reports = { workspace = true }
 lemmy_db_views_private_message = { workspace = true }
 lemmy_db_views_local_image = { workspace = true }
-lemmy_db_views_inbox_combined = { workspace = true }
-lemmy_db_views_person_content_combined = { workspace = true }
-lemmy_db_views_person_saved_combined = { workspace = true }
 lemmy_db_views_report_combined = { workspace = true }
-lemmy_db_views_search_combined = { workspace = true }
 lemmy_db_views_site = { workspace = true }
 lemmy_db_views_registration_applications = { workspace = true }
 lemmy_db_schema = { workspace = true }
