+++ crates/apub/src/api/list_person_content.rs
@@ -1,15 +1,13 @@
 use super::resolve_person_id_from_id_or_username;
 use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
-use lemmy_api_common::{
-  context::LemmyContext,
-  person::{ListPersonContent, ListPersonContentResponse},
-  utils::check_private_instance,
-};
+use lemmy_api_common::{context::LemmyContext, utils::check_private_instance};
 use lemmy_db_schema::traits::PaginationCursorBuilder;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_db_views_person_content_combined::{
   impls::PersonContentCombinedQuery,
+  ListPersonContent,
+  ListPersonContentResponse,
   PersonContentCombinedView,
 };
 use lemmy_db_views_site::SiteView;
