+++ crates/apub/src/api/search.rs
@@ -3,13 +3,17 @@ use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
 use lemmy_api_common::{
   context::LemmyContext,
-  site::{Search, SearchResponse},
   utils::{check_conflicting_like_filters, check_private_instance},
 };
 use lemmy_apub_objects::objects::community::ApubCommunity;
 use lemmy_db_schema::{source::community::Community, traits::PaginationCursorBuilder};
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_search_combined::{impls::SearchCombinedQuery, SearchCombinedView};
+use lemmy_db_views_search_combined::{
+  impls::SearchCombinedQuery,
+  Search,
+  SearchCombinedView,
+  SearchResponse,
+};
 use lemmy_db_views_site::SiteView;
 use lemmy_utils::error::LemmyResult;
