+++ crates/db_views/person_saved_combined/Cargo.toml
@@ -37,6 +37,7 @@ diesel-async = { workspace = true, optional = true }
 serde = { workspace = true }
 ts-rs = { workspace = true, optional = true }
 i-love-jesus = { workspace = true, optional = true }
+serde_with = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
