+++ crates/db_views/search_combined/Cargo.toml
@@ -39,6 +39,7 @@ diesel-async = { workspace = true, optional = true }
 serde = { workspace = true }
 ts-rs = { workspace = true, optional = true }
 i-love-jesus = { workspace = true, optional = true }
+serde_with = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
